+++ include/roq/cache/order.hpp
@@ -131,21 +131,21 @@ struct ROQ_PUBLIC Order final {
   }
 
   uint16_t stream_id = {};
-  const String<MAX_LENGTH_ACCOUNT> account;
+  const Account account;
   const uint32_t order_id = {};
-  const String<MAX_LENGTH_EXCHANGE> exchange;
-  const String<MAX_LENGTH_SYMBOL> symbol;
+  const Exchange exchange;
+  const Symbol symbol;
   Side side = {};
   PositionEffect position_effect = {};
   double max_show_quantity = NaN;
   OrderType order_type = {};
   TimeInForce time_in_force = {};
   ExecutionInstruction execution_instruction = {};
-  String<MAX_LENGTH_ORDER_TEMPLATE> order_template;
+  OrderTemplate order_template;
   std::chrono::nanoseconds create_time_utc = {};
   std::chrono::nanoseconds update_time_utc = {};
-  String<MAX_LENGTH_EXTERNAL_ACCOUNT> external_account;
-  String<MAX_LENGTH_EXTERNAL_ORDER_ID> external_order_id;
+  ExternalAccount external_account;
+  ExternalOrderId external_order_id;
   OrderStatus status = {};
   double quantity = NaN;
   double price = NaN;
@@ -156,7 +156,7 @@ struct ROQ_PUBLIC Order final {
   double last_traded_quantity = NaN;
   double last_traded_price = NaN;
   Liquidity last_liquidity = {};
-  String<MAX_LENGTH_ROUTING_ID> routing_id;
+  RoutingId routing_id;
   uint32_t max_request_version = {};
   uint32_t max_response_version = {};
   uint32_t max_accepted_version = {};