+++ include/roq/cache/statistics.hpp
@@ -145,8 +145,8 @@ class Statistics final {
   }
 
   uint16_t stream_id = {};
-  String<MAX_LENGTH_EXCHANGE> exchange;
-  String<MAX_LENGTH_SYMBOL> symbol;
+  Exchange exchange;
+  Symbol symbol;
   std::array<roq::Statistics, StatisticsType::count()> statistics;
   std::chrono::nanoseconds exchange_time_utc = {};