+++ include/roq/oms/order.hpp
@@ -11,21 +11,21 @@ namespace oms {
 struct ROQ_PUBLIC Order final {
   uint8_t user_id = {};
   uint16_t stream_id = {};
-  String<MAX_LENGTH_ACCOUNT> account;
+  Account account;
   uint32_t order_id = {};
-  String<MAX_LENGTH_EXCHANGE> exchange;
-  String<MAX_LENGTH_SYMBOL> symbol;
+  Exchange exchange;
+  Symbol symbol;
   Side side = {};
   PositionEffect position_effect = {};
   double max_show_quantity = NaN;
   OrderType order_type = {};
   TimeInForce time_in_force = {};
   ExecutionInstruction execution_instruction = {};
-  String<MAX_LENGTH_ORDER_TEMPLATE> order_template;
+  OrderTemplate order_template;
   std::chrono::nanoseconds create_time_utc = {};
   std::chrono::nanoseconds update_time_utc = {};
-  String<MAX_LENGTH_EXTERNAL_ACCOUNT> external_account;
-  String<MAX_LENGTH_EXTERNAL_ORDER_ID> external_order_id;
+  Account external_account;
+  ExternalOrderId external_order_id;
   OrderStatus status = {};
   double quantity = NaN;
   double price = NaN;
@@ -36,7 +36,7 @@ struct ROQ_PUBLIC Order final {
   double last_traded_price = NaN;
   double last_traded_quantity = NaN;
   Liquidity last_liquidity = {};
-  String<MAX_LENGTH_ROUTING_ID> routing_id;
+  RoutingId routing_id;
   uint32_t max_request_version = {};
   uint32_t max_response_version = {};
   uint32_t max_accepted_version = {};